examples: create nidcpower helpers functions
