+++ examples/nidcpower_source_dc_voltage/NIDCPowerSourceDCVoltage.seq
@@ -5434,7 +5434,7 @@
 																	<PythonCall classname='CPythonCall'>
 																		<subprops>
 																			<UseAdapterSettingsForInterpreterSession classname='Bool'>
-																				<value>false</value>
+																				<value>true</value>
 																			</UseAdapterSettingsForInterpreterSession>
 																			<InterpreterSessionScope classname='Num'>
 																				<value>2</value>
